==> notebooks/2)Preprocessing_news_data_kaggle.ipynb <==
## Preprocessing 

# All imports
import pandas as pd
import pandas as pd
import re
import string
import nltk
import random
from datetime import datetime, timedelta
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
from nltk.stem import WordNetLemmatizer
import warnings
warnings.filterwarnings('ignore')

#### Preprocessing function

def preprocess_text(text):
 
    # Convert text to lowercase
    text = text.lower()

    # Remove numbers and special characters
    text = re.sub(r'\d+', '', text)
    text = re.sub(r'[^a-zA-Z0-9\s\']', '', text)
    text = text.translate(str.maketrans('', '', string.punctuation))

    # Tokenize the text into words
    words = word_tokenize(text)

    # Remove stopwords
    stop_words = set(stopwords.words('english'))
    words = [word for word in words if word not in stop_words]

    # Lemmatize words
    lemmatizer = WordNetLemmatizer()
    words = [lemmatizer.lemmatize(word) for word in words]

    # Join the preprocessed words back into a single string
    preprocessed_text = ' '.join(words)

    return preprocessed_text

## 1) Kaggle Data (2013 - 2017)

News=pd.read_csv('crypto_news_parsed_2013-2017_train.csv') 

News1 = News.drop(['url','html','author','source'], axis=1)

News1.head()

News1 = News1.dropna() #getting rid of full empty rows
print(News1.count())

# Apply the preprocess_text function to the DataFrame column
News1['preprocessed_title'] = News1['title'].apply(preprocess_text)

# Apply the preprocess_text function to the DataFrame column
News1['preprocessed_text'] = News1['text'].apply(preprocess_text)

#combining both
processed = News1['preprocessed_text'] + News1['preprocessed_title']
# Create a new DataFrame for further process
processed_kaggle = pd.DataFrame(processed)
processed_kaggle.columns = ['news']

start_date = datetime(2013, 1, 1)  
end_date = datetime(2017, 12, 31)
processed_kaggle['date'] = News1['year'].apply(lambda year: start_date + timedelta(days=random.randint(0, (end_date - start_date).days)))

processed_kaggle

## 2) Scraped crypto News

Crnews=pd.read_csv('crypto_news.csv')
Crnews1 = Crnews.drop(['URL'], axis=1)
Crnews1

Crnews1 = Crnews1.dropna() #getting rid of full empty rows
print(Crnews1.count())

# Apply the preprocess_text function to the DataFrame column
Crnews1['Title'] = Crnews1['Title'].apply(preprocess_text)

Crnews1.columns = ['news','date']
# Convert the 'date' column to datetime objects
Crnews1['date'] = pd.to_datetime(Crnews1['date'], format='%Y-%m-%d %H:%M:%S')

# Now, you can use .dt.strftime() on the 'date' column
Crnews1['date'] = Crnews1['date'].dt.strftime('%Y-%m-%d')
Crnews1['date'] = pd.to_datetime(Crnews1['date'])

Crnews1

# Merge the DataFrames 
merged_df = pd.concat([processed_kaggle, Crnews1], ignore_index=True)

merged_df.to_csv('preprocessed_crypto_data.csv', index=False)
