==> notebooks/3)Labeling Using Cryptobert.ipynb <==
import pandas as pd
from transformers import TextClassificationPipeline, AutoModelForSequenceClassification, AutoTokenizer

News=pd.read_csv('preprocessed_crypto_data.csv') 
News

# converting in to a list to give input to model
text_list = News["news"].tolist()

#using cryptobert model to label the data
model_name = "ElKulako/cryptobert"
tokenizer = AutoTokenizer.from_pretrained(model_name, use_fast=True)
model = AutoModelForSequenceClassification.from_pretrained(model_name, num_labels = 3)
pipe = TextClassificationPipeline(model=model, tokenizer=tokenizer, max_length=64, truncation=True, padding = 'max_length')
result = pipe(text_list)

News['Sentiment'] = [item['label'] for item in result]
News

News.to_csv('labelled_crypto_data.csv', index=False)
