==> notebooks/4) Final_Model_Creation.ipynb <==
import pandas as pd
from imblearn.over_sampling import SMOTE
import warnings
import spacy
import numpy as np
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import classification_report
from sklearn.model_selection import GridSearchCV
from collections import Counter
from sklearn.preprocessing import Normalizer
warnings.filterwarnings('ignore')

df = pd.read_csv("labelled_crypto_data.csv")
#label encoding
df['label'] = df['Sentiment'].map({
    'Bullish' : 0, 
    'Neutral': 1, 
    'Bearish': 2, 
})
df = df.drop(["date","Sentiment"],axis = 1)
df.head()

# Vectorisation using Word2Vec
nlp = spacy.load("en_core_web_lg")
df['vector'] = df['news'].apply(lambda text: nlp(text).vector)

df.head()
df.to_csv("vectorised_output.csv",index = False)

# Oversampling using smote
X = np.stack(df['vector'])
y = df['label']
# Check the class distribution before oversampling
print("Class distribution before oversampling:", Counter(df["label"]))

smote = SMOTE(random_state=42)
X_re,Y_re = smote.fit_resample(X, y)

# Check the class distribution after oversampling
print("Class distribution after oversampling:", Counter(Y_re))

# Hyperparameter tuning and using cross validation
model = RandomForestClassifier()
params = {
    'n_estimators': [10, 50, 100],
    'criterion': ['gini', 'entropy'],
    'max_depth': [None, 10, 20, 30],  
}
grid_search = GridSearchCV(model, params, cv=5, return_train_score=False,n_jobs=-1)
grid_search.fit(X_re, Y_re)
print('params',grid_search.best_params_)
print('score',grid_search.best_score_)

# saving the best model to pickle
import pickle
model = grid_search.best_estimator_
model_filename = 'crypto_model.pkl'
with open(model_filename, 'wb') as file:
    pickle.dump(model, file)
